requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==4.2.29
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
Pillow==11.3.0
packaging==26.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.11.0
